Separação do código em pastas, add função de visão da senha, criação do componente de inputs
