+++ view/src/App.js
@@ -1,11 +1,11 @@
 import React, { useEffect, useState } from 'react';
 import axios from 'axios';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
-import SobreNos from './components/SobreNos';
-import LoginPage from './components/LoginPage'
-import Register from './components/Register'
-import Forgot from './components/Forgot'
-import Grafico from './components/Grafico.js'
+import SobreNos from './components/AboutUs/SobreNos';
+import LoginPage from './components/LoginPage/LoginPage'
+import Register from './components/RegisterPage/Register'
+import Forgot from './components/Forgot/Forgot'
+import Grafico from './components/Grafico/Grafico'
 function App() {
     const [data, setData] = useState(null);
