--- view/src/components/index.html
@@ -1,4 +1,0 @@
-<div>
-    <h1 class = "text-red-600">React App</h1>
-    {data ? <p>{data.message}</p> : <p>Loading...</p>}
-</div>
